+++ dataclients/kubernetes/kube.go
@@ -11,9 +11,9 @@ import (
 	"time"
 
 	log "github.com/sirupsen/logrus"
-	"github.com/zalando/skipper/certregistry"
 	"github.com/zalando/skipper/eskip"
 	"github.com/zalando/skipper/filters"
+	"github.com/zalando/skipper/secrets/certregistry"
 )
 
 const (
